Revision User's Case
I finished the revision on our User's Case. I had to create some news that were missing . Please take a look the way that I implement the case"Iniciar rodada". Because i put the action on a button inside the screen Editar Rodada , but the button doesn't exist until now